requirements.txt
asgiref==3.2.10
authentication==1.1.0
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2==2.8.6
pycryptodome==3.9.8
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
Werkzeug==1.0.1
whitenoise==5.2.0
WTForms==2.3.3
